Add basic level goal BP
Place this BP in your level and a message
will be printed on screen once the required
number of units enter the bounding box.
The threshold defaults to 1.0 meaning all units
need to be inside for the level to end. It can be
set to another value per instance of the blueprint.